Puts serialization in separate classes
