+++ lib/macaroons/raw_macaroon.rb
@@ -4,30 +4,31 @@ require 'rbnacl'
 
 require 'macaroons/caveat'
 require 'macaroons/utils'
+require 'macaroons/serializers/binary'
 
 module Macaroons
   class RawMacaroon
-    PACKET_PREFIX_LENGTH = 4
 
-    def initialize(key: nil, identifier: nil, location: nil, serialized: nil)
-      if key.nil? && identifier.nil? && location.nil? && serialized.nil?
-        raise ArgumentError, 'Must provide either (key, id, location), or serialized.'
-      elsif (key.nil? || identifier.nil? || location.nil?) and serialized.nil?
+    def initialize(key: nil, identifier: nil, location: nil)
+      if key.nil? || identifier.nil? || location.nil?
         raise ArgumentError, 'Must provide all three: (key, id, location)'
       end
 
       @key = key
       @identifier = identifier
       @location = location
-      @signature = create_initial_macaroon_signature(key, identifier) unless serialized
+      @signature = create_initial_macaroon_signature(key, identifier)
       @caveats = []
-      deserialize(serialized) unless serialized.nil?
+    end
+
+    def self.from_binary(serialized: serialized)
+      Macaroons::BinarySerializer.new().deserialize(serialized)
     end
 
     attr_reader :identifier
     attr_reader :key
     attr_reader :location
-    attr_reader :caveats
+    attr_accessor :caveats
     attr_accessor :signature
 
     def signature
@@ -52,23 +53,7 @@ module Macaroons
     end
 
     def serialize
-      combined = packetize('location', @location)
-      combined += packetize('identifier', @identifier)
-
-      for caveat in @caveats
-        combined += packetize('cid', caveat.caveat_id)
-
-        if caveat.verification_id and caveat.caveat_location
-          combined += packetize('vid', caveat.verification_id)
-          combined += packetize('cl', caveat.caveat_location)
-        end
-      end
-
-      combined += packetize(
-        'signature',
-        Utils.unhexlify(self.signature)
-      )
-      Base64.urlsafe_encode64(combined)
+      Macaroons::BinarySerializer.new().serialize(self)
     end
 
     def prepare_for_request(macaroon)
@@ -86,60 +71,6 @@ module Macaroons
     end
 
     private
-
-    def deserialize(serialized)
-      @caveats = []
-
-      decoded = Base64.urlsafe_decode64(serialized)
-
-      index = 0
-
-      while index < decoded.length
-        packet_length = decoded[index..(index + PACKET_PREFIX_LENGTH - 1)].to_i(16)
-        stripped_packet = decoded[index + PACKET_PREFIX_LENGTH..(index + packet_length - 2)]
-
-        key, value = depacketize(stripped_packet)
-
-        case key
-        when 'location'
-          @location = value
-        when 'identifier'
-          @identifier = value
-        when 'cid'
-          @caveats << Caveat.new(value)
-        when 'vid'
-          @caveats[-1].verification_id = value
-        when 'cl'
-          @caveats[-1].caveat_location = value
-        when 'signature'
-          @signature = value
-        else
-          raise KeyError, 'Invalid key in binary macaroon. Macaroon may be corrupted.'
-        end
-
-        index = index + packet_length
-      end
-    end
-
-    def packetize(key, data)
-      # The 2 covers the space and the newline
-      packet_size = PACKET_PREFIX_LENGTH + 2 + key.length + data.length
-      if packet_size > 65535
-        # Due to packet structure, length of packet must be less than 0xFFFF
-        raise ArgumentError, 'Data is too long for a binary packet.'
-      end
-      packet_size_hex = packet_size.to_s(16)
-      header = packet_size_hex.to_s.rjust(4, '0')
-      packet_content = "#{key} #{data}\n"
-      packet = "#{header}#{packet_content}"
-      packet
-    end
-
-    def depacketize(packet)
-      key = packet.split(" ")[0]
-      value = packet[key.length + 1..-1]
-      [key, value]
-    end
 
     def create_initial_macaroon_signature(key, identifier)
       derived_key = Utils.generate_derived_key(key)
