+++ lib/macaroons/serializers/base.rb
@@ -1,0 +1,13 @@
+module Macaroons
+  class BaseSerializer
+
+    def serialize(macaroon)
+      raise NotImplementedError
+    end
+
+    def deserialize(serialized)
+      raise NotImplementedError
+    end
+
+  end
+end
